Describe aquí los cambios que hiciste

+++ controladores/eliminarTurno.php
@@ -4,23 +4,33 @@ include_once '../bd/conexion.php';
 
 if ($_SERVER['REQUEST_METHOD'] === 'POST' && isset($_GET['id_turno'])) {
     $id_turno = $_GET['id_turno'];
+    $errores = [];
 
-    $sql = $conexion->prepare("DELETE FROM turnos WHERE id = :id");
-    $sql->bindParam(':id', $id_turno);
-    $sql->execute();
+    // ------Validaciones-----
 
-    if ($sql == true) {
-        $_SESSION['mensaje'] = "Turno eliminado correctamente."; // Mensaje de éxito en sesión
+    if(empty($id_turno)) {
+        $errores[] = "ID de turno es obligatorio.";
+    } elseif (!ctype_digit($id_turno)) {
+        $errores[] = "ID de turno inválido.";
+    }
+
+    if(empty($errores)){
+        $sql = $conexion->prepare("DELETE FROM turnos WHERE id = :id");
+        $sql->bindParam(':id', $id_turno, PDO::PARAM_INT);
+        $sql->execute();
+
+        $_SESSION['mensaje'] = "Turno eliminado exitosamente.";
         header("Location: ../vista_consulta_general.php");
         exit();
     } else {
-        $_SESSION['mensaje'] = "Error al eliminar el turno.";
+        $_SESSION['mensaje'] = $errores;
+        header("Location: ../vista_consulta_general.php");
         exit();
     }
-}else{
-    echo "Error al eliminar el turno";
+
+} else {
+    $_SESSION['mensaje'] = "Error al procesar la solicitud.";
+    header("Location: ../vista_consulta_general.php");
     exit();
 }
-
-
 ?>
